# swagger.yaml
openapi: 3.0.0
info:
  title: Sistema de Gestión de Solicitudes
  version: "1.0.0"
  description: |
    API REST robusta para la gestión de solicitudes de soporte técnico.  
    Permite la administración de usuarios, roles, solicitudes, historial de cambios y reportes.  
    Seguridad avanzada con JWT, CORS, rate limiting y validaciones.  
    Documentación exhaustiva para todos los endpoints, parámetros, respuestas y errores posibles.
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Usuario:
      type: object
      description: Usuario del sistema con credenciales y rol.
      properties:
        id:
          type: integer
          description: Identificador único del usuario.
          example: 1
        nombre:
          type: string
          description: Nombre completo del usuario.
          example: "María Cliente"
        email:
          type: string
          description: Correo electrónico único.
          example: "cliente1@test.com"
        rol:
          type: string
          description: Rol del usuario en el sistema.
          enum: [CLIENTE, SOPORTE, ADMIN]
          example: "CLIENTE"
      required: [nombre, email, rol]
...
    Error:
      type: object
      description: Respuesta de error estándar.
      properties:
        error:
          type: string
          example: "Error en el servidor"
        code:
          type: integer
          example: 500
security:
  - bearerAuth: []
...
paths:
...
  /usuarios:
    post:
      tags: [Usuarios]
      summary: Registrar nuevo usuario
      description: |
        Permite registrar un usuario nuevo en el sistema.  
        El rol puede ser CLIENTE, SOPORTE o ADMIN.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                  example: "Pedro Cliente"
                email:
                  type: string
                  example: "pedro@test.com"
                password:
                  type: string
                  example: "pedro123"
                rol:
                  type: string
                  enum: [CLIENTE, SOPORTE, ADMIN]
                  example: "CLIENTE"
      responses:
        '201':
          description: Usuario creado exitosamente.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Usuario creado"
                  usuario:
                    $ref: '#/components/schemas/Usuario'
        '400':
          description: Error de validación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email ya registrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Error en el servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /usuarios/{id}:
    get:
      tags: [Usuarios]
      summary: Obtener usuario por ID
      description: |
        Retorna los datos de un usuario específico por su ID.  
        Solo accesible por admin o el propio usuario.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID del usuario.
      responses:
        '200':
          description: Usuario encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '403':
          description: Sin permisos para ver el usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Usuario no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Error en el servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags: [Usuarios]
      summary: Actualizar usuario por ID
      description: |
        Permite actualizar los datos de un usuario.  
        Solo el propio usuario o admin pueden modificar.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID del usuario.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                  example: "Pedro Actualizado"
                email:
                  type: string
                  example: "pedro2@test.com"
                password:
                  type: string
                  example: "nuevo123"
                rol:
                  type: string
                  enum: [CLIENTE, SOPORTE, ADMIN]
                  example: "SOPORTE"
      responses:
        '200':
          description: Usuario actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '400':
          description: Error de validación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Sin permisos para actualizar el usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Usuario no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Error en el servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Usuarios]
      summary: Eliminar usuario por ID
      description: |
        Permite eliminar un usuario del sistema.  
        Solo admin puede eliminar usuarios.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID del usuario.
      responses:
        '204':
          description: Usuario eliminado exitosamente.
        '403':
          description: Sin permisos para eliminar el usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Usuario no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Error en el servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...